backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
requests
